Testes de servico do livro prontos

+++ DonaRosangela/DonaRosangela.App/Features/Books/BookService.cs
@@ -1,4 +1,5 @@
-﻿using DonaRosangela.Domain.Features.Books;
+﻿using DonaRosangela.Domain.Exceptions;
+using DonaRosangela.Domain.Features.Books;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -18,16 +19,19 @@ namespace DonaRosangela.App.Features.Books
 
         public Book Add(Book obj)
         {
+            obj.Validade();
             return _repository.Add(obj);
         }
 
         public void Delete(Book obj)
         {
+            if (obj.Id == 0) throw new InvalidIdException();
             _repository.Delete(obj.Id);
         }
 
         public Book Get(Book obj)
         {
+            if (obj.Id == 0) throw new InvalidIdException();
             return _repository.Get(obj.Id);
         }
 
@@ -38,6 +42,7 @@ namespace DonaRosangela.App.Features.Books
 
         public Book Update(Book obj)
         {
+            obj.Validade();
             return _repository.Update(obj);
         }
     }
